trait hoisted up to roo
