+++ src/add-traits.ts
