+++ src/merge-traits.test.ts
