+++ src/merge-traits.ts
